classe para ler csv

+++ Musica.java
@@ -16,7 +16,7 @@ public class Musica {
   protected String name, key;
   protected ArrayList<String> artists;
   protected int id, duration_ms, explicit;
-  protected float tempo;
+  protected double tempo;
   protected Date release_date;
 
   // constructors
@@ -28,7 +28,7 @@ public class Musica {
     this.release_date = null;
   }
 
-  public Musica (int id, String key, String name, ArrayList<String> artists, int duration_ms, int explicit, float tempo, Date release_date) {
+  public Musica (int id, String key, String name, ArrayList<String> artists, int duration_ms, int explicit, double tempo, Date release_date) {
     this.id = id;
     this.key = key;
     this.name = name;
@@ -46,7 +46,7 @@ public class Musica {
   public void setArtists(ArrayList<String> artists) { this.artists = artists; }
   public void setDuration_ms(int duration_ms) { this.duration_ms = duration_ms; }
   public void setExplicit(int explicit) { this.explicit = explicit; }
-  public void setTempo(float tempo) { this.tempo = tempo; }
+  public void setTempo(double tempo) { this.tempo = tempo; }
   public void setRelease_date(Date release_date) { this.release_date = release_date; }
 
   // gets
@@ -56,7 +56,7 @@ public class Musica {
   public ArrayList<String> getArtists() { return artists; }
   public int getDuration_ms() { return duration_ms; }
   public int getExplicit() { return explicit; }
-  public float getTempo() { return tempo; }
+  public double getTempo() { return tempo; }
   public Date getRelease_date() { return release_date; }
 
   // ----------------------------------- fim constructor -----------------------------------
@@ -78,7 +78,7 @@ public class Musica {
     }
     dos.write(duration_ms);
     dos.write(explicit);
-    dos.writeFloat(tempo);
+    dos.writeDouble(tempo);
     dos.writeUTF(release_date_string);
     return baos.toByteArray();
   }
@@ -98,7 +98,7 @@ public class Musica {
 
     duration_ms = dis.read();
     explicit = dis.read();
-    tempo = dis.readFloat();
+    tempo = dis.readDouble();
 
     // converter string para data
     String release_date_string = dis.readUTF();
